Quash compiler warnings in pbl.c.

+++ pbl.c
@@ -83,7 +83,7 @@ char * pblx_get_cstring (CFStringRef data) {
 #include <execinfo.h>
 #include <stdlib.h>
 #define CHECK(sub,var) \
-    fprintf (stderr, "Debug %s - %s returned %ld\n", ROUTINE, sub, var); \
+    fprintf (stderr, "Debug %s - %s returned %ld\n", ROUTINE, sub, ( long ) var); \
     if (var) goto cleanup;
 #define LOG_C(var,dta) { \
     if ((dta) == NULL) { \
@@ -729,7 +729,7 @@ int main (int argc, char **argv) {
 		    data[size] = '\0';
 		    printf( "data: '%s'\n", data );
 		    printf( "size: %lu\n", size );
-		    printf( "flags: %#lx\n", flags );
+		    printf( "flags: %#lx\n", ( unsigned long ) flags );
 		}
 		CFRelease (pbref);
 	    }
@@ -752,7 +752,7 @@ int main (int argc, char **argv) {
 		for ( inx = 0; inx < num_resp; inx++ ) {
 		    printf( "\nid: %lu\n", resp[inx].id );
 		    printf( "flavor: %s\n", resp[inx].flavor );
-		    printf( "flavor flags: %#lx\n", resp[inx].flags );
+		    printf( "flavor flags: %#lx\n", ( unsigned long ) resp[inx].flags );
 		    if ( resp[inx].data != NULL ) {
 			printf( "data: %s\n", resp[inx].data );
 			printf( "size: %lu\n", resp[inx].size );
@@ -771,7 +771,7 @@ int main (int argc, char **argv) {
 	help ();
     }
     if (stat) {
-	fprintf (stderr, "Error - Status = %li\n", stat);
+	fprintf (stderr, "Error - Status = %li\n", ( long ) stat);
     }
     return stat ? 1 : 0;
 }
